Tune queries by checking for relevant nulls
Check for not null ELM when retrieving from the export tables, as it's very unlikely an export would be created with the ELM XML.

Check for not null severe error on measure_xml since that's the only column having its charset adjusted from that table. Similar for reference in measure_details_reference.