Stop removing the first Translator error
The `TranslatorOptionsRemover` was not actively checking for a `translatorOptions` Node following the assumption that the first Node in the `annotation` json array would always contain `translatorOptions` as the first object.

This no longer a valid assumption, there are cases where the Translator does not include the `translatorOptions` node, and it needs to be explicitly verified before removal.

Reviewing the CQl-to-ELM Translator docs, I cannot determine what controls the inclusion of `translatorOptions` in its output.