Exclude FHIRHelpers during unused library check
When building the list of unused libraries, ignore FHIRHelpers.

FHIRHelpers is used by the Translator to handle conversion between the FHIR and CQL System models. A common example is when comparing an element on a FHIR Resource that has a required binding (i.e. Encounter.status ~ 'finished'), the left side of the comparison is a FHIR type, FHIR.EncounterStatus, while the right side is System.String. FHIRHelpers is by the Translator to convert the left side to System.String for a consistent comparison.

Thus, FHIRHelpers can be used without being explicitly referenced in a library's definitions or functions, which would normally make it "unused" and eligible for optional removal at versioning.